refactor: fulltext 적용 부분 리팩토링(#107)
