+++ src/main/java/ku_rum/backend/global/dataInit/FulltextIndexInitializer.java
@@ -1,0 +1,46 @@
+package ku_rum.backend.global.dataInit;
+
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.boot.context.event.ApplicationReadyEvent;
+import org.springframework.context.event.EventListener;
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.stereotype.Component;
+
+@Slf4j
+@Component
+public class FulltextIndexInitializer {
+
+    private final JdbcTemplate jdbcTemplate;
+
+    public FulltextIndexInitializer(JdbcTemplate jdbcTemplate) {
+        this.jdbcTemplate = jdbcTemplate;
+    }
+
+    @EventListener(ApplicationReadyEvent.class)
+    public void init() {
+        try {
+            insert_fulltext_index("building", "name_fulltext_index");
+            insert_fulltext_index("category", "name_fulltext_index");
+        } catch (Exception e) {
+            log.error("❌ FULLTEXT INDEX 추가 실패: {}", e.getMessage(), e);
+        }
+    }
+
+    private void insert_fulltext_index(String tableName, String indexName) {
+        if (!isIndexExists(tableName,indexName)){
+            jdbcTemplate.execute(
+                    "ALTER TABLE tableName ADD FULLTEXT INDEX indexName (name) WITH PARSER ngram"
+            );
+        }
+    }
+
+    private boolean isIndexExists(String tableName, String indexName) {
+        String sql = "SELECT COUNT(*) FROM information_schema.statistics " +
+                "WHERE table_schema = DATABASE() " +
+                "AND table_name = ? " +
+                "AND index_name = ?";
+
+        return jdbcTemplate.queryForObject(sql, Integer.class, tableName, indexName) > 0;
+    }
+}
+
